Define data contract for Observation objects
By assigning DataContract and DataMember attributes to Object, fields respectively, we can serialize Observation objects into JSON with a DataContractSerializer. This allows for saving of Observation objects as strings in Preferences (nice as supported across platforms), and loading Observation data from Preferences and re-converting into objects.

These objects are used to populate values of UI elements which appear to user in MainPage  view.